requirements.txt
# requirements.txt (배포용)
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.38
alembic==1.15.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.4.0
passlib==1.7.4
email_validator==2.2.0
python-multipart==0.0.20
requests
openai==1.70.0
tensorflow==2.19.0
scikit-learn==1.6.1
celery[redis]
redis